Green, Orange, & Purple are moving (Small tweaks)
Updated the mixed patron trio with animations much like the last three so they can walk and idle and randomly change idles. I also went in and changed the blue, yellow, and red organization. Now the script and animation components are on the actual model instead of on the whole prefab